# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChapterCreate": {
        "properties": {
          "chapter_no": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter No"
          },
          "title": {
            "default": "",
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          }
        },
        "title": "ChapterCreate",
        "type": "object"
      },
...
      "ErrorAction": {
        "enum": [
          "retry",
          "reload",
          "wait",
          "view_diff",
          "open_resync",
          "unlock_vault",
          "open_provider_settings",
          "edit_instruction",
          "change_model",
          "adjust_budget"
        ],
        "title": "ErrorAction",
        "type": "string"
      },
      "ErrorCode": {
        "enum": [
          "VALIDATION",
          "REVISION_CONFLICT",
          "WORK_BLOCKED",
          "WORK_BUSY_QUEUED",
          "WORK_ACTIVE_JOB",
          "CHAPTER_RANGE_CONFLICT",
          "CHAPTER_READ_ONLY",
          "CANDIDATE_NOT_READY",
          "CANDIDATE_CLOSED",
          "CHAPTER_IS_BASE",
          "STATE_SNAPSHOT_MISSING",
          "CHAPTER_EMPTY",
          "VAULT_LOCKED",
          "VAULT_PASSWORD_INVALID",
          "VAULT_UNLOCK_THROTTLED",
          "VAULT_ALREADY_EXISTS",
          "VAULT_NOT_FOUND",
          "VAULT_CORRUPT",
          "SECRET_MISSING",
          "PROVIDER_AUTH",
          "PROVIDER_UNREACHABLE",
          "PROVIDER_SERVER_ERROR",
          "PROVIDER_RATE_LIMIT",
          "PROVIDER_IN_USE",
          "PROVIDER_REFUSAL",
          "OUTPUT_TRUNCATED",
          "STRUCTURED_OUTPUT_INVALID",
          "BUDGET_EXCEEDED",
          "UNAUTHORIZED",
          "FORBIDDEN_HOST",
          "FORBIDDEN_ORIGIN",
          "NOT_FOUND",
          "IDEMPOTENCY_CONFLICT",
          "DB_BUSY",
          "BACKEND_SHUTTING_DOWN",
          "INTERNAL"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorAction"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ReorderIn": {
        "properties": {
          "chapter_ids": {
            "items": {
              "type": "string"
            },
            "title": "Chapter Ids",
            "type": "array"
          }
        },
        "required": [
          "chapter_ids"
        ],
        "title": "ReorderIn",
        "type": "object"
      },
      "ReplaceIn": {
        "properties": {
          "doc_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Json"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "ReplaceIn",
        "type": "object"
      },
...
      "RestoreIn": {
        "properties": {
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "RestoreIn",
        "type": "object"
      },
...
      "SnapshotIn": {
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "reason": {
            "default": "manual_snapshot",
            "enum": [
              "leave_chapter",
              "idle",
              "manual_snapshot",
              "before_ai_accept",
              "after_ai_accept",
              "before_restore",
              "full_replace"
            ],
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "SnapshotIn",
        "type": "object"
      },
...
      "WorkingCopyIn": {
        "properties": {
          "base_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Revision Id"
          },
          "client_seq": {
            "default": 0,
            "minimum": 0.0,
            "title": "Client Seq",
            "type": "integer"
          },
          "client_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Session Id"
          },
          "doc_json": {
            "additionalProperties": true,
            "title": "Doc Json",
            "type": "object"
          }
        },
        "required": [
          "doc_json"
        ],
        "title": "WorkingCopyIn",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "WriteStoryApp backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chapters/{chapter_id}": {
      "delete": {
        "operationId": "delete_chapter",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Delete Chapter",
        "tags": [
          "chapters"
        ]
      },
      "get": {
        "operationId": "get_chapter",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Chapter",
        "tags": [
          "chapters"
        ]
      },
      "put": {
        "operationId": "replace_chapter",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Replace Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
...
    "/v1/chapters/{chapter_id}/revisions": {
      "get": {
        "operationId": "list_chapter_revisions",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List Chapter Revisions",
        "tags": [
          "chapters"
        ]
      }
    },
    "/v1/chapters/{chapter_id}/revisions/{before_id}/diff/{after_id}": {
      "get": {
        "operationId": "diff_chapter_revisions",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "before_id",
            "required": true,
            "schema": {
              "title": "Before Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "after_id",
            "required": true,
            "schema": {
              "title": "After Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Diff Chapter Revisions",
        "tags": [
          "chapters"
        ]
      }
    },
    "/v1/chapters/{chapter_id}/revisions/{revision_id}": {
      "get": {
        "operationId": "get_chapter_revision",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Chapter Revision",
        "tags": [
          "chapters"
        ]
      }
    },
    "/v1/chapters/{chapter_id}/revisions/{revision_id}/restore": {
      "post": {
        "operationId": "restore_chapter_revision",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Restore Chapter Revision",
        "tags": [
          "chapters"
        ]
      }
    },
...
    "/v1/chapters/{chapter_id}/snapshot": {
      "post": {
        "operationId": "snapshot_chapter",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Snapshot Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
...
    "/v1/chapters/{chapter_id}/working-copy": {
      "get": {
        "operationId": "get_working_copy",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Working Copy",
        "tags": [
          "chapters"
        ]
      },
      "put": {
        "operationId": "save_working_copy",
        "parameters": [
          {
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "title": "Chapter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingCopyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Save Working Copy",
        "tags": [
          "chapters"
        ]
      }
    },
...
    "/v1/works/{work_id}/chapters": {
      "get": {
        "operationId": "list_chapters",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List Chapters",
        "tags": [
          "chapters"
        ]
      },
      "post": {
        "operationId": "create_chapter",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Create Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
    "/v1/works/{work_id}/chapters/reorder": {
      "post": {
        "operationId": "reorder_chapters",
        "parameters": [
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Reorder Chapters",
        "tags": [
          "chapters"
        ]
      }
    },
...
  }
}